@@ -17,50 +17,46 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^18.2.2",
-    "@angular/cdk": "^18.2.2",
-    "@angular/common": "^18.2.2",
-    "@angular/compiler": "^18.2.2",
-    "@angular/core": "^18.2.2",
-    "@angular/forms": "^18.2.2",
-    "@angular/material": "^18.2.2",
-    "@angular/platform-browser": "^18.2.2",
-    "@angular/platform-browser-dynamic": "^18.2.2",
-    "@angular/router": "^18.2.2",
-    "@ngrx/effects": "^18.0.2",
-    "@ngrx/operators": "^18.0.0",
-    "@ngrx/store": "^18.0.2",
-    "@ngrx/store-devtools": "^18.0.2",
+    "@angular/animations": "^20.1.7",
+    "@angular/cdk": "^20.1.6",
+    "@angular/common": "^20.1.7",
+    "@angular/compiler": "^20.1.7",
+    "@angular/core": "^20.1.7",
+    "@angular/forms": "^20.1.7",
+    "@angular/material": "^20.1.6",
+    "@angular/platform-browser": "^20.1.7",
+    "@angular/platform-browser-dynamic": "^20.1.7",
+    "@angular/router": "^20.1.7",
+    "@ngrx/effects": "^20.0.0",
+    "@ngrx/operators": "^20.0.0",
+    "@ngrx/store": "^20.0.0",
+    "@ngrx/store-devtools": "^20.0.0",
     "d3": "^7.9.0",
     "rxjs": "~7.8.1",
     "tslib": "^2.6.2",
-    "zone.js": "^0.14.10"
+    "zone.js": "^0.15.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^18.2.2",
-    "@angular-eslint/builder": "^18.3.0",
-    "@angular-eslint/eslint-plugin": "^18.3.0",
-    "@angular-eslint/eslint-plugin-template": "^18.3.0",
-    "@angular-eslint/schematics": "^18.3.0",
-    "@angular-eslint/template-parser": "^18.3.0",
-    "@angular/cli": "^18.2.2",
-    "@angular/compiler-cli": "^18.2.2",
-    "@eslint/eslintrc": "^3.1.0",
+    "@angular-devkit/build-angular": "^20.1.6",
+    "@angular/cli": "^20.1.6",
+    "@angular/compiler-cli": "^20.1.7",
+    "@eslint/eslintrc": "^3.3.1",
     "@eslint/js": "^9.9.1",
     "@types/d3": "^7.4.3",
-    "@types/jasmine": "^5.1.4",
-    "@web/test-runner": "^0.18.2",
-    "eslint": "^9.9.1",
-    "eslint-plugin-unused-imports": "^4.1.3",
-    "jasmine-core": "^5.1.2",
+    "@types/jasmine": "^5.1.9",
+    "@web/test-runner": "^0.20.2",
+    "angular-eslint": "^20.1.1",
+    "eslint": "^9.28.0",
+    "eslint-plugin-unused-imports": "^4.1.4",
+    "jasmine-core": "^5.9.0",
     "karma-chrome-launcher": "^3.2.0",
     "karma-coverage": "^2.2.1",
     "karma-jasmine": "^5.1.0",
     "karma-jasmine-html-reporter": "^2.1.0",
-    "prettier": "^3.3.0",
+    "prettier": "^3.6.2",
     "ts-node": "^10.9.2",
-    "typescript": "^5.4.5",
-    "typescript-eslint": "^8.3.0",
+    "typescript": "^5.8.3",
+    "typescript-eslint": "^8.39.1",
     "webpack-bundle-analyzer": "^4.10.2"
   }
 }
